Conversion/Hybrid/Divergence minorities
Converting faith, creatinge a hybrid culture, and creating a divergent culture now all leave minorities.
For faiths, when you and your capital convert, it leaves no minority if righteous, a small one if astray, and a large one if hostile or evil.
For hybrids and divergences, it leaves a large minority of the old parent culture in counties that switch to the new culture.