add special high partition variant
Adds a new succession law, Ultimogenitary Partition, that is basically High Partition but with a youngest-first succession order, and a doctrine to Chaldean and Nasrani and Tradition to Syriac that makes the succession available. In order to add the tradition to your culture you need to either be Syriac, a culture descended from Syriac, a Chaldean or Nasrani Christian, or a culture descended from another culture that already has the tradition.